Fix testFindClasses test
This test is actually failing in master currently, as well as the 5.6 branch if that change is cherry-picked into that branch.
The concern with this test failing was that the order of the returned classes was important, but given that the updated order is the current behaviour of both master and the 5.6 branch it can't actually be a problem to just switch the order in the test. Presumably this relates to update in the Symfony finder since this test was created originally.